pyproject.toml
[project]
name = "lyrics-video-sync"
version = "0.1.0"
description = "异步歌词语义混剪后端"
requires-python = ">=3.11"
authors = [
  { name = "twelve_labs", email = "dev@twelvelabs.local" },
]
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "httpx>=0.27.0",
  "twelvelabs>=0.1.7",
  "sqlmodel>=0.0.16",
  "asyncpg>=0.29.0",
  "redis>=5.0.1",
  "arq>=0.25.0",
  "pydub>=0.25.1",
  "python-ffmpeg>=2.0.4",
  "openai-whisper>=20231117",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "python-dotenv>=1.0.0",
  "pydantic-settings>=2.6.1",
  "anyio>=4.3.0",
  "structlog>=23.2.0",
  "aiosqlite>=0.19.0",
  "minio>=7.2.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "respx>=0.21.1",
  "pytest-subprocess>=1.5.0",
  "ruff>=0.5.0",
  "mypy>=1.9.0",
  "types-redis",
  "types-PyYAML",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = 3.11
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
plugins = []
